# lib/api-zod/src/generated/types/heritageExplanation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HeritageExplanationFeaturesItem } from './heritageExplanationFeaturesItem';

export interface HeritageExplanation {
  title: string;
  summary: string;
  features: HeritageExplanationFeaturesItem[];
  note: string;
}