# lib/api-zod/src/generated/types/heritageExplanationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HeritageExplanationInputLanguage } from './heritageExplanationInputLanguage';
import type { HeritageExplanationInputMimeType } from './heritageExplanationInputMimeType';

export interface HeritageExplanationInput {
  /** @minLength 20 */
  image_base64: string;
  mime_type: HeritageExplanationInputMimeType;
  language: HeritageExplanationInputLanguage;
}