# lib/api-zod/src/generated/types/heritageSite.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HeritageSite {
  name: string;
  region: string;
  period: string;
  accent: string;
}